Autu's sparse A part stored as vector<double*> G
the Aij can be obtained by
G[i][j]

For sake of memory efficienct, since Atut is only read and would not be
increases. G[i]=new double[max element in row i]();

The only problem is that we would not get G[i]'s size easily. Since we
do not need to use it. That's find for now.